Support records writes and index merge updates
Adds write() to Storage to allow for a record to be written to the data
cf, and then a merge of the record's ID to the index (in the idx cf).

Need to understand whether it is possible to write a record but then for
the merge op to go missing. Can we batch a write+merge?